account for ITL filename changes, and correct for processing of *.txt files; refactor VendorFitsTranslator classes
