+++ harnessed_jobs/vendorIngest/v0/validator_vendorIngest.py
@@ -1,4 +1,6 @@
 #!/usr/bin/env python
+"vendorIngest harnessed job validator script."
+from __future__ import absolute_import, print_function
 import os
 import sys
 import shutil
@@ -8,22 +10,30 @@ import ConfigParser
 import numpy as np
 import lcatr.schema
 import siteUtils
-from vendorFitsTranslators import ItlFitsTranslator, e2vFitsTranslator
+from ItlFitsTranslator import ItlFitsTranslator
+from e2vFitsTranslator import e2vFitsTranslator
 
 def validate(schema, **kwds):
+    "More compact call to lcatr.schema.valid"
     return lcatr.schema.valid(lcatr.schema.get(schema), **kwds)
 
 class VendorResults(object):
-    amps = range(1, 17)
+    """
+    Base class to process vendor results.
+    """
     qe_band_passes = dict([('u', (321, 391)),
                            ('g', (402, 552)),
                            ('r', (552, 691)),
                            ('i', (691, 818)),
                            ('z', (818, 922)),
                            ('y', (930, 1070))])
+
     def __init__(self):
-        pass
+        "Constructor."
+        self._amps = range(1, 17)
+
     def run_all(self):
+        "Run all of the methods for each test type."
         results = []
         failures = OrderedDict()
         analyses = ('fe55_analysis', 'read_noise', 'bright_defects',
@@ -33,18 +43,18 @@ class VendorResults(object):
             try:
                 exec('my_results = self.%s()' % analysis)
                 results.extend(my_results)
-            except Exception, eobj:
+            except Exception as eobj:
                 failures[analysis] = eobj
         if failures:
-            print
-            print "Failed to extract vendor results for the following:"
+            print("\nFailed to extract vendor results for the following:")
             for analysis, eobj in failures.items():
-                print "%s: %s, %s" % (analysis, type(eobj), eobj)
-            print
+                print("%s: %s, %s" % (analysis, type(eobj), eobj))
+            print("")
         sys.stdout.flush()
         return results
 
 class ItlResults(VendorResults):
+    "Class to process the ITL results."
     file_end_mapping = {'fe55_analysis' : 'fe55.txt',
                         'bright_defects' : 'brightdefects.txt',
                         'dark_defects' : 'darkdefects.txt',
@@ -58,16 +68,22 @@ class ItlResults(VendorResults):
                         'qe_analysis' : 'qe.txt',
                         'metrology' : 'metrology.txt'}
     def __init__(self, rootdir):
+        """
+        Constructor.
+        rootdir = Top level directory containing all of the results files.
+        """
         super(ItlResults, self).__init__()
-        command = 'find %(rootdir)s/ -name \*.txt -print' % locals()
+        command = 'find %s/ -name \*.txt -print' % rootdir
         text_files = subprocess.check_output(command, shell=True).split()
-        print "Found ITL results files:"
+        print("Found ITL results files:")
         for item in text_files:
-            print "  ", item
+            print("  ", item)
         self.inverse_mapping = dict([(os.path.basename(path), path) for path
                                      in text_files])
         self._configs = {}
+
     def __getitem__(self, key):
+        "Provide a dict-like interface to each .ini style subsection."
         if not self._configs.has_key(key):
             self._configs[key] = ConfigParser.ConfigParser()
             file_ending = self.file_end_mapping[key]
@@ -76,12 +92,21 @@ class ItlResults(VendorResults):
                     target = os.path.basename(item)
                     break
             self._configs[key].read(self.inverse_mapping[target])
+            # Set the number of amplifiers for this txt file based on
+            # the [Info] section.
+            info = dict(self._configs[key].items('Info'))
+            try:
+                self._amps = range(1, int(info['numchans']) + 1)
+            except KeyError:
+                pass
         return self._configs[key]
+
     def fe55_analysis(self):
+        "Process the Fe55 analysis results."
         job = 'fe55_analysis'
         gains = dict(self[job].items('SystemGain'))
         results = []
-        for amp in self.amps:
+        for amp in self._amps:
             ext = '%02i' % (amp - 1)
             amp_catalog = dict(self[job].items('Events Channel %s' % ext))
             results.append(validate(job, amp=amp,
@@ -89,19 +114,21 @@ class ItlResults(VendorResults):
                                     gain_error=0,
                                     psf_sigma=amp_catalog['meansigma']))
         return results
+
     def read_noise(self):
+        "Process the read noise results."
         job = 'read_noise'
         noise = dict(self[job].items('ReadNoise'))
         results = []
         system_noise_data = {}
-        for amp in self.amps:
+        for amp in self._amps:
             ext = '%02i' % (amp - 1)
             read_noise = float(noise['readnoise_%s' % ext])
             try:
                 system_noise = float(noise['systemnoisecorrection_%s' % ext])
                 system_noise_data[amp] = system_noise
             except KeyError:
-                system_noise = 0
+                system_noise = 0.
             total_noise = np.sqrt(read_noise**2 + system_noise**2)
             results.append(validate(job, amp=amp,
                                     read_noise=read_noise,
@@ -114,54 +141,62 @@ class ItlResults(VendorResults):
                 for amp, system_noise in system_noise_data.items():
                     output.write('  %i        %f\n' % (amp, system_noise))
         return results
+
     def bright_defects(self):
+        "Process the bright defects results."
         job = 'bright_defects'
         defects = dict(self[job].items('BrightRejection'))
-        print """
+        print("""
         For the bright defects results, ITL only provides the total
         number of rejected pixels, so set all of these to be
         bright_pixels in amp 1 and set everything else to zero.
-        """
+        """)
         total_rejected = int(defects['brightrejectedpixels'])
         results = [validate(job, amp=1, bright_pixels=total_rejected,
                             bright_columns=0)]
-        for amp in self.amps:
+        for amp in self._amps:
             if amp == 1:
                 continue
             results.append(validate(job, amp=amp,
                                     bright_pixels=0, bright_columns=0))
         return results
+
     def dark_defects(self):
+        "Process the dark defects results."
         job = 'dark_defects'
         defects = dict(self[job].items('DarkRejection'))
-        print """
+        print("""
         For the dark defects results, ITL only provides the total
         number of rejected pixels, so set all of these to be
         dark_pixels in amp 1 and set everything else to zero.
-        """
+        """)
         total_rejected = int(defects['darkrejectedpixels'])
         results = [validate(job, amp=1, dark_pixels=total_rejected,
                             dark_columns=0)]
-        for amp in self.amps:
+        for amp in self._amps:
             if amp == 1:
                 continue
             results.append(validate(job, amp=amp,
                                     dark_pixels=0, dark_columns=0))
         return results
+
     def traps(self):
+        "Process the traps results."
         job = 'traps'
         results = []
-        for amp in self.amps:
+        for amp in self._amps:
             results.append(validate(job, amp=amp, num_traps=-1))
         return results
+
     def dark_current(self):
+        "Process the dark current results."
         job = 'dark_current'
         dc = dict(self[job].items('DarkSignal'))
-        print """
+        print("""
         For the dark current results, ITL only provides CCD-wide
         numbers for the current at any given percentile, so set all
         amps to have this same value.
-        """
+        """)
         # Need to loop through DarkFrac# entries to find the 95th
         # percentile value.
         index = None
@@ -171,12 +206,14 @@ class ItlResults(VendorResults):
         if index is not None:
             dc_value = float(dc['darkrate'+index])
         else:
-            dc_value = -1  # ugly sentinel value
+            dc_value = -1.  # ugly sentinel value
         results = []
-        for amp in self.amps:
+        for amp in self._amps:
             results.append(validate(job, amp=amp, dark_current_95CL=dc_value))
         return results
+
     def cte(self):
+        "Process the CTE results."
         job = 'cte_low'
         scte_low = dict(self[job].items('HCTE'))
         pcte_low = dict(self[job].items('VCTE'))
@@ -184,7 +221,7 @@ class ItlResults(VendorResults):
         scte_high = dict(self[job].items('HCTE'))
         pcte_high = dict(self[job].items('VCTE'))
         results = []
-        for amp in self.amps:
+        for amp in self._amps:
             ext = '%02i' % (amp - 1)
             results.append(validate('cte_vendorIngest', amp=amp,
                                     cti_low_serial=1.-float(scte_low['hcte_%s' % ext]),
@@ -192,7 +229,9 @@ class ItlResults(VendorResults):
                                     cti_high_serial=1.-float(scte_high['hcte_%s' % ext]),
                                     cti_high_parallel=1.-float(pcte_high['vcte_%s' % ext])))
         return results
+
     def prnu(self):
+        "Process the PRNU results."
         job = 'prnu'
         prnu = dict(self[job].items('PRNU'))
         results = []
@@ -206,26 +245,32 @@ class ItlResults(VendorResults):
                                     pixel_stdev=prnu_percent,
                                     pixel_mean=100.))
         return results
+
     def flat_pairs(self):
+        "Process the flat pairs results."
         job = 'flat_pairs'
         residuals = dict(self[job].items('Residuals'))
-        max_frac_devs = dict((amp, 0) for amp in self.amps)
-        full_wells = dict((amp, 0) for amp in self.amps)
+        max_frac_devs = dict((amp, 0) for amp in self._amps)
+        full_wells = dict((amp, 0) for amp in self._amps)
         for key, value in residuals.items():
             if key.startswith('residuals'):
                 devs = [float(x.strip())/100. for x in value.split()]
-                for amp, dev in zip(self.amps, devs):
+                for amp, dev in zip(self._amps, devs):
                     if np.abs(dev) > max_frac_devs[amp]:
                         max_frac_devs[amp] = np.abs(dev)
         results = []
-        for amp in self.amps:
+        for amp in self._amps:
             results.append(validate(job, amp=amp,
                                     full_well=full_wells[amp],
                                     max_frac_dev=max_frac_devs[amp]))
         return results
+
     def ptc(self):
+        "Process the PTC results."
         return []
+
     def qe_analysis(self):
+        "Process the QE results."
         job = 'qe_analysis'
         qe_data = dict(self[job].items('QE'))
         qe_results = dict((band, []) for band in self.qe_band_passes)
@@ -242,7 +287,9 @@ class ItlResults(VendorResults):
             results.append(validate(job, band=band,
                                     QE=np.average(qe_results[band])))
         return results
+
     def metrology(self):
+        "Process the metrology results."
         job = 'metrology'
         test_results = {}
         try:
@@ -270,9 +317,17 @@ class ItlResults(VendorResults):
         return results
 
 class e2vResults(VendorResults):
+    "Class to process e2v CCD results."
     def __init__(self, rootdir):
+        """
+        Constructor.
+        rootdir = Top level directory containing all of the results files.
+        """
+        super(e2vResults, self).__init__()
         self.rootdir = rootdir.replace(' ', '\ ')
+
     def _csv_data(self, *args, **kwds):
+        "Method to extract per amp data from csv file."
         amp_data = {}
         subpath = os.path.join(*args)
         command = 'find %s/ -name %s -print' % (self.rootdir, subpath)
@@ -289,7 +344,9 @@ class e2vResults(VendorResults):
             amp = int(tokens[0])
             amp_data[amp] = [x.strip() for x in tokens[1:]]
         return amp_data.items()
+
     def fe55_analysis(self):
+        "Process Fe55 results."
         job = 'fe55_analysis'
         gains = {}
         psf_sigmas = {}
@@ -298,12 +355,14 @@ class e2vResults(VendorResults):
         for amp, tokens in self._csv_data('\*PSF\*_Summary\*.csv'):
             psf_sigmas[amp] = float(tokens[0])
         results = []
-        for amp in self.amps:
+        for amp in self._amps:
             results.append(validate(job, amp=amp, gain=gains[amp],
                                     gain_error=0,
                                     psf_sigma=psf_sigmas[amp]))
         return results
+
     def read_noise(self):
+        "Process the read noise results."
         job = 'read_noise'
         results = []
         for amp, tokens in self._csv_data('\*Noise\*Multiple\*Samples\*Summary\*.csv'):
@@ -314,7 +373,9 @@ class e2vResults(VendorResults):
                                     system_noise=system_noise,
                                     total_noise=total_noise))
         return results
+
     def bright_defects(self):
+        "Process the bright defects results."
         job = 'bright_defects'
         results = []
         for amp, tokens in self._csv_data('\*Darkness_Summary\*.csv'):
@@ -323,7 +384,9 @@ class e2vResults(VendorResults):
             results.append(validate(job, amp=amp, bright_pixels=bright_pixels,
                                     bright_columns=bright_columns))
         return results
+
     def dark_defects(self):
+        "Process the dark defects results."
         job = 'dark_defects'
         results = []
         for amp, tokens in self._csv_data('\*PRDefs_Summary\*.csv'):
@@ -332,14 +395,18 @@ class e2vResults(VendorResults):
             results.append(validate(job, amp=amp, dark_pixels=dark_pixels,
                                     dark_columns=dark_columns))
         return results
+
     def traps(self):
+        "Process the traps results."
         job = 'traps'
         results = []
         for amp, tokens in self._csv_data('\*TrapsPP_Summary\*.csv'):
             num_traps = int(tokens[0])
             results.append(validate(job, amp=amp, num_traps=num_traps))
         return results
+
     def dark_current(self):
+        "Process the dark current results."
         job = 'dark_current'
         results = []
         for amp, tokens in self._csv_data('\*Darkness_Summary\*.csv'):
@@ -347,7 +414,9 @@ class e2vResults(VendorResults):
             results.append(validate(job, amp=amp,
                                     dark_current_95CL=dark_current))
         return results
+
     def cte(self):
+        "Process the CTE results."
         job = 'cte_vendorIngest'
         results = []
         scti_low, pcti_low, scti_high, pcti_high = {}, {}, {}, {}
@@ -357,14 +426,16 @@ class e2vResults(VendorResults):
         for amp, tokens in self._csv_data('\*CTE\*Optical\*High_Summary\*.csv'):
             pcti_high[amp] = 1. - float(tokens[0])
             scti_high[amp] = 1. - float(tokens[1])
-        for amp in self.amps:
+        for amp in self._amps:
             results.append(validate(job, amp=amp,
                                     cti_low_serial=scti_low[amp],
                                     cti_low_parallel=pcti_low[amp],
                                     cti_high_serial=scti_high[amp],
                                     cti_high_parallel=pcti_high[amp]))
         return results
+
     def prnu(self):
+        "Process the PRNU results."
         job = 'prnu'
         results = []
         for wl, tokens in self._csv_data('\*PRNU_Summary\*.csv',
@@ -373,7 +444,9 @@ class e2vResults(VendorResults):
             results.append(validate(job, wavelength=wl,
                                     pixel_stdev=prnu_percent, pixel_mean=100))
         return results
+
     def flat_pairs(self):
+        "Process the flat pairs results."
         job = 'flat_pairs'
         results = []
         for amp, tokens in self._csv_data('\*FWC\*Multiple\*Image\*Summary\*.csv'):
@@ -383,9 +456,13 @@ class e2vResults(VendorResults):
                                     full_well=full_well,
                                     max_frac_dev=max_frac_dev))
         return results
+
     def ptc(self):
+        "Process the PTC results."
         return []
+
     def qe_analysis(self):
+        "Process the QE results."
         job = 'qe_analysis'
         subpath = '\*QE_Summary\*.csv'
         command = 'find %s/ -name %s -print' % (self.rootdir, subpath)
@@ -420,7 +497,9 @@ class e2vResults(VendorResults):
             results.append(validate(job, band=band,
                                     QE=np.average(qe_results[band])))
         return results
+
     def metrology(self):
+        "Process the metrology results."
         return []
 
 def ITL_metrology_files(rootdir, expected_num=1):
@@ -487,7 +566,7 @@ if __name__ == '__main__':
     lsstnum = siteUtils.getUnitId()
 
     vendorDataDir = os.readlink('vendorData')
-    print 'Vendor data location:', vendorDataDir
+    print('Vendor data location:', vendorDataDir)
 
     if siteUtils.getCcdVendor() == 'ITL':
         vendor = ItlResults(vendorDataDir)
